Move builder into repo
